wibble: Automated memory management
I integrated the memorymanager into the Parser.  To ensure that all generated
components are properly memory managed, I set their constructors as private.

I had to setup special wrapper templates to redefine the methods as public, for
the tests to be able to instantiate the classes they try to test.

In order to remove a complicated dependency chain where all classes would have
to know about MemoryManager and MemoryManager would have to know about all the
classes, I created a common 'Component' base class, so that MemoryManager can
know how to delete the various components it is passed.

The only inconvenience I can see so far is the use of -> because many of the
things are pointer.  I guess that the only way to avoid that is to switch to
Ruby :-/

Anyway, this commandline parser is approaching perfection, please nitpick as
much as possible, I can't wait to use it in all my applications! :)