wibble: untangle MorphBase into MorphInterface + implementation interface
Untangled the MorphBase class, which was previously contained within
the implementation interface via inheritance. Now there is
MorphInterface, which is direct base of MorphImpl and implementation
interface which is also direct base of MorphImpl. However, the two
interfaces are not in inheritance relation like they used to be. This
simplifies the case of multiple inheritance for morph implementations
-- now a single class can be part of multiple amorph sets.