Rework layout of Blobs used for graph node storage.
To make it possible to use variable-length vertex representation, we put the
algorithm-specific "extension" in front of the actual node data. This
"extension" data is bound to be of fixed size anyway. The graph generator can
now request appropriately sized chunk of memory for its nodes (although it
obviously needs to know the right size before using that memory).