OWCTY counterexamples, part one: finding an accepting vertex lying on a cycle.
This is arguably the most tricky part. The code is now able to do the
counterexample search, locating a vertex that lies on an accepting cycle. The
rest of full counterexample support is to print a trace to the initial state
(as in reachability) and then to find (and store, using parent pointers) a path
from the selected vertex back to itself.