Rework counterexamples to collect traces on the master node.
We previously relied on mpiexec to collect stderrs of participating processes
and print them in order. This was a rather unreliable interim solution. The
current code will synchronise after every step and do all state printing on the
master. There is currently no observable behaviour change (other than the
bugfix). This also opens up the path to process the counterexample
(e.g. produce a sequence of edges instead of a sequence of states) or write it
to a file (instead of just stderr).