LLVM: Rework the (p)thread mechanics somewhat.
We now track thread IDs inside the LLVM interpreter in divine. This allows
checking for cross-thread mutex unlocks and the human-readable thread
description gained thread IDs, which makes it easier to keep track of threads
as they come and go in traces. In addition, we now have a partial
pthreads-compatible mutex API, including a prototype recursive mutex
implementation.