Refactor algorithm instantiation and init code.
The template selection and instantation has moved into the library, the
interface between the runner and the algorithms now uses a couple of virtual
methods. The library interface is provided by divine/select.h. Additionally, we
now distribute the Meta instance to all worker threads, which simplifies
algorithm implementation somewhat.