LLVM: New Pointer representation.
I have worked out a reasonable compromise for exact pointer tracking and we are
aiming for that now, instead of a conservative approach. This will enable a
fully canonic heap and a cheap store observability analysis while reducing
state snapshot size. The drawbacks are that pointer values (even when cast to
integers) must always be stored aligned (not a serious problem) and that
integer load instructions will incur slightly more complex static analysis.