toolkit: Deprecate ABI-exposed Pool counters with overflow tendencies.
There is a pair of per-group counters in Pool that are declared as "int" and
tend to overflow on x86-64 systems, especially for big fixed-state-size
models. This could in theory trigger a rather subtle NULL dereference bug when
total overflow'd to 0. Other than that, assert-enabled builds have had a
penchant for crashing on big models.