Toolkit: Rework Stores adding Vertex and VertexId to fix memory management.
Due to fact that not all stores are in fact storing full states any more
(HcStore, TreeCompressedStore) it is necessary to solve memory management 
which is currently expecting all nodes to be permanent on worker who created
them.

To solve this problem Vertex and VertexId Store dependent types are introduced.
Vertex is intended to be sent in queues, it should contains both full node and
permanent identifier. VertexId is permanent identifier to be stored in
extension and used for CE generation. It must be possible to fetch extension
based only on VertexId (on worker owning this Id). Also it must be possible 
get ownwer of VertexId and two VertexIds are equal iff they represent
equal vertices, it must be possible to test equality on any worker.

Direct access to underliing table from algorithms by means of operator[]
was dropped by making TableUtils private. Instead iterator support for store
was added -- store now supports begin() and end() methods giving iterators 
to VertexIds of uderliing table which allows iterating over all VertexIds 
using C++11 range-based for loop. This way of iteration should be sufficient
in most cases since it allows direct access to saved Extensions.

Also TableUtils class was split do separate file (divine/graph/tableutils.h)
and it is no more predecessor of stores but their member.

Shared store is so far left to be resolved later.