toolkit: Fix malformed assert in NTreeHashSet::Leaf().
For some reason NTreeHashSet was asserting that leafs have size at least 2.
However, this is not required by any part of NTreeHashSet and in fact it can
happen that generator splits node in way that leafs have size 1. (It currently
happens for all states with size mod 32 = 1 with exception of Timed generator.)