LLVM: Remove an over-ambitious assert on address validity.
Since we use char *dereference of ValueRefs for the purpose of bound checking,
we need to be able to handle the cases where the ValueRef points inside a valid
range, even though it doesn't fit entirely. This is not really a problem, since
all (potentially dangerous) dereferences need to be guarded by a bounds check
anyway.