toolkit: More compact cell specialization for Blob in SharedHashSet.
We can use tag of Blob (alias Lake::Pointer) to store part of hash not used for
index calculation. Using this we can drop dedicated hash field from Cell
as collisions are likely to be resolved by 16bit hash fragment in tag.

The upside is Cell is now only 8bytes long (instead of 16) making table
half the size for same capacity. Also no Cell locking is necessary now.
The downside is that hash has to be recomputed when resizing table.

Performance is to be experimentally assessed.