utility: Add reporting to SQL using ODBC.
It is now possible to save report from verification into SQL database
using arbitrary database supporting ODBC.
The report queries database for list of columns in given table, and saves
one row containing just columns wich are both in database and in report
(mangled names are used -- lowercase with undersore instead of nonalphanumerical
characters).
If database does not exits report will attemt to create it with all columns
that can currently be in report.