toolkit: Make operator bool explicit everywhere it is declared.
It is bad idea to have implicit operator bool as it can result in misbehaved
comparison binary operators (and other bool taking functions).
On the other hand having explicit operator bool has virtually no downsides
as if and ! imposes explicit bool context (that is value with explicit
operator bool in if or after ! is implicitly converted to bool).