LLVM: Work around an aliasing problem with MemoryBits.
For some reason, clang 3.3 optimizes out too much of MemoryBits::get(), making
it sometimes return 0 when it shouldn't. This happens even when only
type-punning through unions and void pointers (which should be well-defined),
and only when get() is inlined into its caller. We might be unwittingly
invoking some undefined behaviour somewhere, or it might be a clang/llvm
optimizer bug.