NIX: Use ld.gold from gcc's binutils.
To avoid clash between libstdc++ in gold and LLVMgold. This is not entirely
good solution as there is no guearantee that this version of gold is
built with the same libstdc++ as LLVMgold, but since LLVM is build usually
with standard gcc this should work.
Note: This works also in case we use self-hosted LLVM based on libc++,
simpli for the fact that LLVMgold does not used libstdc++ in that case
and there is no clash.