LLVM: Avoid ending threads to generate one more state after join.
The previous order of operations was as follows: when pthread_join
was called it signalled thread (using flag) to go to detached state,
causing it to release all resources *after* pthread_join returned.
However such a behaviour resulted in the fact that numbers of states
were multiplied by number combinations of threads which were joined
earlier, as freeing of those threads could have been postponed
indefinitely. This resulted it hight state space overhead
if any nontrivial work was done after some thread was joined.
To mitigate this problem, pthread_join is now forced to wait
for _pthread_entry to release all resources; as this is the last
action of the (ending) thread, it should not generate any other state
at least if reductions are enabled.
The recently added thread_termination.cpp test can be used as example
(compiling with -DCNT=# to give number of threads to run), before
the change, it was not possible to verify more than 8 threads in short
time, now even 32 are ok.