toolkit: Invalidate cell before it is moved & multiple fixes of atomic cells.
If cell is invalidated after it was moved it can happen, that given
cell will first be detected empty in rehashing procedure, then written
and then invalidated, resulting in invalidated but not moved cell.
Also bug when writting to cell could clear invalidate flag was eliminated
(invalidate now waits for writes to end).
And finally in case of AtomicCell invalidating empty cell caused no
change to the cell and therefore it could have happen that
other thread inserted into this invalidated cell.