algorithm: Fix race condition in parallel nested-dfs.
I a rare case outer exploration of parallel nested-dfs can indicate that work
is done earlier than inner visitor starts processing nodes. This results in
situations when loop is not found and whole state space is explored by outer
visitor.

The buggy execution:
1.  outer thread found accepting state but did to push it to communicating queue
    yet line < 175
2.  inner is running, checks queue (it is empty) and is stopped by scheduler
    around line 78
3.  outer pushes into queue, finishes exploration ad sets finished = true
    (l. 127)
4.  outer runs test on line 79 and finishes without poping queue

It should be sufficient to check whether queue is empty and stop just in case
it is empty (or if cycle was found, therefore valid = false).