toolkit: Avoid 4B of overhead in tree compression (for larger states).
We can avoid 4B of overhead per state by not storing degree of root of tree.
The downside is that the degree must now be at least one (there are no flat
states any more). This affects small states where there will be additional
overhead of 4B, but this is not as significant since small states usually imply
small number of states and three compression should not be used in such case
anyway.