runtime: For uncaught C++ exceptions, __vm_choose whether to unwind the stack.
This simulates both behaviours allowed by the C++ standard, which says that it
is implementation defined whether stack is unwound (and destructors run) if no
catch handler is found for an exception. However, as Itanium ABI does not
specify this, we use this only for C++ exceptions (based on the exception
class).